feat(workflow) add the complete workflow for banquepopulaire.fr
In this iteration was added a complete version of the workflow for monitoring the banquepopulaire.fr. With this addition we updated the folder structure to separate the images and the configuration getters basically creating /settings and /images. The settings now are pulled dinamically although they are still declared for autocomplete reasons.