feat(report) add report and video functionality
With the addition of video and the creation of an PDF report. The idea was to basically join two consequences of the constant screenshot framing to give tools the user can have more details of what is happening. There were changes in the file directory to represent that basically now this app is divided in four parts REVISEUR, REPORT, VIDEO and SETTINGS