feat(selenium) adds selenim to workflow
With this commit we started using selenium we can put the operation on background. And in case of error the program will create a report and video besides its logs. The idea it is to use selenium to crawl the website and then compare snapshots to those expected by Lackey